+++ UserOperaction.h
@@ -10,10 +10,10 @@ class UserOperaction
     int userId;
 	int operationId;
 	int date;
-    float amount;
+    double amount;
 
 public:
-    UserOperaction(int userId = 0,int operationId = 0,int date = 0, float amount = 0.0)
+    UserOperaction(int userId = 0,int operationId = 0,int date = 0, double amount = 0.0)
     {
         this -> userId = userId;
 		this -> operationId = operationId;
@@ -24,12 +24,12 @@ public:
     void setUserId(int newUserId);
     void setOperationId(int newOperationId);
     void setDate(int newDate);
-    void setAmount(float newAmount);
+    void setAmount(double newAmount);
 
     int getUserId();
     int getOperationId();
 	int getDate();
-	float getAmount();
+	double getAmount();
 
 };
